Added a counter that counts the attempts that all threads combined needed to find the password.
Known issue:
generated_password (password) and user_password get compared by reviewing each char from the first to the last position as often as the user_password is long. This means that the program will think the generated_password (password) "pascalaadad" is the same as the user_password "pascal". This may result in the program finding a password that isn't the password the user entered